05.09

+++ src/main/java/project/web/data/domain/CityPicture.java
@@ -14,7 +14,7 @@ public class CityPicture {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "cpic_num")
-    private Integer cPic_num;
+    private Long cPic_num;
 
     @Id
     @Column(name = "c_num")
